@@ -5,7 +5,7 @@
   "main": "utils/seed.js",
   "scripts": {
     "start": "node bin/www",
-    "test": "jest"
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js"
   },
   "keywords": [],
   "author": "Christopher Bartos, Charles-Alexandre Bouchard, Oni Picotte",
@@ -13,11 +13,11 @@
   "dependencies": {
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
-    "eslint": "^8.51.0",
     "express": "^4.18.2",
     "mongodb": "^6.2.0"
   },
   "devDependencies": {
+    "eslint": "^8.52.0",
     "jest": "^29.7.0",
     "supertest": "^6.3.3"
   }
