@@ -1,4 +1,4 @@
-const DB = require('../db/DB');
+const DB = require('../db/db');
 
 (async () => {
   let db;
@@ -7,9 +7,11 @@ const DB = require('../db/DB');
     // dbname is cluster0 in my case
     await db.connect('dataset', 'stocks');
     const num = await db.deleteMany({});
+    // eslint-disable-next-line no-console
     console.log(`Deleted ${num} stocks`);
   } catch (e) {
     console.error('could not delete');
+    // eslint-disable-next-line no-console
     console.dir(e);
   } finally {
     if (db) {
